Econ-ARK ISO cleanup and make for econ-ark-tools

+++ Tools/Install/Machines/Scripts/Methods-ISO/finish_modified-for-econ-ark.sh
@@ -1,13 +1,13 @@
 #!/bin/bash
+# Update everything 
 sudo apt -y update && sudo apt -y upgrade
+# Install anaconda3 for all users 
 # Adapted from http://askubuntu.com/questions/505919/how-to-install-anaconda-on-ubuntu
 
-sudo ls  # make sure we have root permission from the outset
-scriptDir="$(dirname "`realpath $0`")"
+sudo ls  > /dev/null # make sure we have root permission
 
-if [ -e /tmp/Anaconda ]; then # delete any prior install
-    sudo rm -Rf /tmp/Anaconda
-fi
+# Put in /tmp directory
+[[ -e /tmp/Anaconda ]] && sudo rm -Rf /tmp/Anaconda # delete any prior install
 
 mkdir /tmp/Anaconda ; cd /tmp/Anaconda
 
@@ -23,30 +23,35 @@ cmd="sudo rm -Rf /usr/local/anaconda3 ; chmod a+x /tmp/Anaconda/$ANACONDAURL ; /
 echo "$cmd"
 eval "$cmd"
 
+# Add to default enviroment path so that everyone can find it
 addToPath='export PATH=/usr/local/anaconda3/bin:$PATH'
+echo "$addToPath"
 eval "$addToPath"
 sudo chmod u+w /etc/environment
 sudo sed -e 's\/usr/local/sbin:\/usr/local/anaconda3/bin:/usr/local/sbin:\g' /etc/environment > /tmp/environment
-sudo sed -e 's\/usr/local/anaconda3/bin:/usr/local/anaconda3/bin\/usr/local/anaconda3/bin\g' /tmp/environment > /tmp/environment2 # eliminate any duplicates
+
+# eliminate any duplicates which may exist if the script has been run more than once
+sudo sed -e 's\/usr/local/anaconda3/bin:/usr/local/anaconda3/bin\/usr/local/anaconda3/bin\g' /tmp/environment > /tmp/environment2
+
 sudo mv /tmp/environment2 /etc/environment # Weird permissions issue prevents direct redirect into /etc/environment
-sudo chmod u-w /etc/environment
+sudo chmod u-w /etc/environment # Restore secure permissions for environment
 
 if [ ! -e /etc/sudoers.d/anaconda3 ]; then # Modify secure path so that anaconda commands will work with sudo
     sudo mkdir -p /etc/sudoers.d
     sudo echo 'Defaults secure_path="/usr/local/anaconda3/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/snap/bin:/bin"' | sudo tee /etc/sudoers.d/anaconda3
 fi
 
-echo "$PATH"
-source /etc/environment
+source /etc/environment  # Get the new environment
 
 conda update --yes conda
 conda update --yes anaconda
 
+# Add some final common tools
 conda install --yes -c anaconda scipy
-conda install --yes -c anaconda pyopengl # Otherwise you get an error "Segmentation fault (core dumped)"
+conda install --yes -c anaconda pyopengl # Otherwise you get an error "Segmentation fault (core dumped)" on some Ubuntu machines
+conda install --yes -c conda-forge jupyter_contrib_nbextensions
 
-$scriptDir/Anaconda-jupyter_contrib_nbextensions.sh
-#/Methods/Tools/Config/tool/jupytext/default.sh
+# Get default packages for Econ-ARK machine
 sudo apt -y install git bash-completion xsel cifs-utils openssh-server nautilus-share xclip texlive-full emacs auctex
 
 #Download and extract HARK, REMARK, DemARK from GitHUB repository
@@ -61,7 +66,8 @@ git clone https://github.com/econ-ark/DemARK.git
 # https://askubuntu.com/questions/499070/install-virtualbox-guest-addition-terminal
 
 sudo apt -y install build-essential module-assistant virtualbox-guest-dkms virtualbox-guest-utils virtualbox-guest-x11
-mkdir -p /home/$myuser/GitHub/econ-ark ; ln -s /usr/local/share/GitHub/econ-ark /home/$myuser/GitHub/econ-ark
+mkdir -p /home/econ-ark/GitHub/econ-ark ; ln -s /usr/local/share/GitHub/econ-ark /home/econ-ark/GitHub/econ-ark
+chown -Rf econ-ark:econ-ark /usr/local/share/GitHub/econ-ark # Make it be owned by econ-ark user 
 
-Finished automatic installations.  Rebooting.
+echo Finished automatic installations.  Rebooting.
 reboot
